Refactor helper functions and more
* Move helper functions from SyncHelper to utils.py
* Obtain default exporter port constant from config.py
* Refactor parse_metrics to simplify logic
* Change test_config_collector_enabled so that it passes only when all
  enabled collectors for the hardware are provided
